+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Feature
 
-* [#3](https://github.com/ryz310/redis-object-daily-counter/pull/3) Support time zone ([@ryz310](https://github.com/ryz310))
-* [#4](https://github.com/ryz310/redis-object-daily-counter/pull/4) Add daily counter family ([@ryz310](https://github.com/ryz310))
+* [#3](https://github.com/ryz310/redis-objects-daily-counter/pull/3) Support time zone ([@ryz310](https://github.com/ryz310))
+* [#4](https://github.com/ryz310/redis-objects-daily-counter/pull/4) Add daily counter family ([@ryz310](https://github.com/ryz310))
     * `annual_counter`
     * `monthly_counter`
     * `weekly_counter`
@@ -14,7 +14,7 @@
 
 ### Misc
 
-* [#2](https://github.com/ryz310/redis-object-daily-counter/pull/2) Install circle ci ([@ryz310](https://github.com/ryz310))
+* [#2](https://github.com/ryz310/redis-objects-daily-counter/pull/2) Install circle ci ([@ryz310](https://github.com/ryz310))
 
 ## v0.1.0 (Sep 12, 2021)
