+++ lib/redis-objects-daily-counter.rb
